refactor(code): consolidate insight and think pages
