--- src/app/(app)/insights/InsightsHero2.tsx
@@ -1,9 +1,0 @@
-export function InsightsHero2() {
-  return (
-    <>
-      <div className="grid grid-cols-[63.25rem] grid-rows-[16.88rem] items-center justify-items-center p-12 text-[9.38rem] leading-none text-neutral-100">
-        <h1>Our latest news and views.</h1>
-      </div>
-    </>
-  );
-}
