+++ src/app/(app)/think/ThinkWonder.tsx
@@ -1,4 +1,4 @@
-export function InsightsWonder() {
+export function ThinkWonder() {
   return (
     <>
       <div>
